length check on sys.argv before looking for hyphen
With more arguments now set via environment variables there's not always a need to run the container with arguments, in which case we get "IndexError: list index out of range". Checking to see if there are any arguments before we try to do something with them seems like a reasonable thing to do.